Add macOS terminal support
