+++ internal/repl/repl.go
@@ -11,9 +11,7 @@ import (
 	"path/filepath"
 	"regexp"
 	"strings"
-	"syscall"
 	"time"
-	"unsafe"
 
 	"github.com/example/grimux/internal/openai"
 	"github.com/example/grimux/internal/tmux"
@@ -95,28 +93,6 @@ func lastCodeBlock(text string) string {
 }
 
 // startRaw puts the terminal into raw mode.
-func startRaw() (*syscall.Termios, error) {
-	fd := int(os.Stdin.Fd())
-	var old syscall.Termios
-	if _, _, err := syscall.Syscall6(syscall.SYS_IOCTL, uintptr(fd), uintptr(syscall.TCGETS), uintptr(unsafe.Pointer(&old)), 0, 0, 0); err != 0 {
-		return nil, err
-	}
-	newState := old
-	newState.Lflag &^= syscall.ICANON | syscall.ECHO
-	if _, _, err := syscall.Syscall6(syscall.SYS_IOCTL, uintptr(fd), uintptr(syscall.TCSETS), uintptr(unsafe.Pointer(&newState)), 0, 0, 0); err != 0 {
-		return nil, err
-	}
-	return &old, nil
-}
-
-// stopRaw restores the terminal to a previous state.
-func stopRaw(state *syscall.Termios) {
-	if state == nil {
-		return
-	}
-	fd := int(os.Stdin.Fd())
-	syscall.Syscall6(syscall.SYS_IOCTL, uintptr(fd), uintptr(syscall.TCSETS), uintptr(unsafe.Pointer(state)), 0, 0, 0)
-}
 
 // Run launches the interactive REPL.
 func Run() error {
